intial commit

+++ Routes/expense.js
@@ -4,6 +4,7 @@ import authenticated from '../middleware/auth.js';
 import { user } from '../Model/UserModel.js';
 import sendFilterdRecord from '../utils/filterRecords.js';
 import { record } from '../Model/recordModel.js';
+import { ApiError } from '../utils/apiError.js';
 
 const expenseRouter = express.Router()
 
@@ -23,15 +24,16 @@ expenseRouter.post('/add',authenticated , async function(req,res){
 
         const data = await Expense.create({title,amount,category,user:req.user,record:recordId,date,month})
 
-        console.log(req.user)
+        if(!data) throw new ApiError('400','','try again')
       
 
-        res.json({data})
+        
+        res.status(201).json({success:true,data})
    
     }catch(e){
 
   
-        console.log(e.message)
+        res.status(500).json(new ApiError(500,'',e.errors))
     }
   
 })
@@ -41,9 +43,10 @@ expenseRouter.get('/getprofile',authenticated,async function(req,res){
     try{
     const userData = await user.findById(req.user._id)
 
-    res.send({data:userData})
+  if(!userData) throw new ApiError(404,'','not found')
+    res.status(200).json({success:true,data:userData})
     }catch(e){
-        res.send({msg:e.message})
+        res.status(404).json(new ApiError(404,'',e.errors))
     }
 })
 
@@ -53,12 +56,14 @@ expenseRouter.get('/currentdata',authenticated,async function(req,res){
     // const skiped = req.query.skip;
     
         try{
-            console.log(req.user._id)
+            
         const userData = await Expense.find({user:id}).skip(req.user.skip).sort({date:1});
     
-        res.send({data:userData})
+            if(!userData) throw new ApiError(404,'','not found')
+
+        res.status(200).json({success:true,data:userData})
         }catch(e){
-            res.send({msg:e.message})
+            res.status(404).json(new ApiError(404,'',e.errors))
         }
     })
 
@@ -68,21 +73,31 @@ expenseRouter.get('/currentdata',authenticated,async function(req,res){
         // const skiped = req.query.skip;
         
             try{
-                console.log(req.user._id)
+                
             const userData = await Expense.find({user:id}).sort({date:1,month:1});
         
-            res.send({data:userData})
+            if(!userData) throw new ApiError(404,'','not found')
+
+            res.json({success:true,data:userData}).status(200)
+
             }catch(e){
-                res.send({msg:e.message})
+                res.status(404).json(new ApiError(404,'',e.errors))
             }
         })
 
 expenseRouter.post('/remove/:id',async function(req,res){
-    const {id} = req.params
+    try {
+        const {id} = req.params
 
-    const entryDeleted = await Expense.findByIdAndDelete({_id:id})
-
-    res.send({msg:'successfully deleted'})
+        if(!id) throw new ApiError(500,'','Reload App')
+    
+        const entryDeleted = await Expense.findByIdAndDelete({_id:id})
+    
+        
+        res.status(200).json({success:true,msg:'deleted successfully'})
+    } catch (error) {
+        res.status(500).json(new ApiError(500,'',e.errors))
+    }
 
    
 })
@@ -93,12 +108,15 @@ expenseRouter.get('/filter',authenticated,async function(req,res){
     const {year,month,date} = req.query
     const id = req.user._id
 
-    console.log(year,month,date)
+    
 
    try {
      if(year) {
          if(!month && !date){
              const recordByYear = await Expense.find({year,user:id}).sort({date:1})
+
+             if(!recordByYear) throw new ApiError(404,'','not found')
+
              sendFilterdRecord(req,res,recordByYear)
              
              }
@@ -107,6 +125,8 @@ expenseRouter.get('/filter',authenticated,async function(req,res){
      if(month) {
          if(!year && !date){
              const recordByMonth = await Expense.find({month,user:id}).sort({date:1})
+
+             if(!recordByMonth) throw new ApiError(404,'','not found')
              sendFilterdRecord(req,res,recordByMonth)
              
              }
@@ -118,6 +138,8 @@ expenseRouter.get('/filter',authenticated,async function(req,res){
              console.log(typeof Number(date))
             
              const recordByDate = await Expense.find({$and:[{date:{$gte:date}},{date:{$lte:31}}],user:id}).sort({date:1})
+
+             if(!recordByData) throw new ApiError(404,'','not found')
              sendFilterdRecord(req,res,recordByDate)
              
              }
@@ -126,6 +148,8 @@ expenseRouter.get('/filter',authenticated,async function(req,res){
      if(year && month ) {
          if(!date){
              const recordByYearMonth = await Expense.find({year,month,user:id}).sort({date:1})
+
+             if(!recordByYearMonth) throw new ApiError(404,'','not found')
              sendFilterdRecord(req,res,recordByYearMonth)
              }
      } 
@@ -133,11 +157,12 @@ expenseRouter.get('/filter',authenticated,async function(req,res){
      if(year && month && date){
          
          const recordByYearMonthDate = await Expense.find({$and:[{date:{$gte:date}},{date:{$lte:31}}],year,month,user:id}).sort({date:1,year:-1,month:-1})
+         if(!recordByYearMonthDate) throw new ApiError(404,'','not found')
          sendFilterdRecord(req,res,recordByYearMonthDate)   
      }
  
    } catch (error) {
-    console.log(error.message)
+    res.status(404).json(new ApiError(404,'',error.errors))
 
     
    }
